register always the plugin
