+++ packages/richtext-lexical/src/features/blocks/client/plugin/index.tsx
@@ -1,15 +1,8 @@
 'use client'
 
-import { copyToClipboard } from '@lexical/clipboard'
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext.js'
-import {
-  $insertNodeToNearestRoot,
-  $wrapNodeInElement,
-  mergeRegister,
-  objectKlassEquals,
-} from '@lexical/utils'
+import { $insertNodeToNearestRoot, $wrapNodeInElement, mergeRegister } from '@lexical/utils'
 import { formatDrawerSlug, useEditDepth } from '@payloadcms/ui'
-import ObjectID from 'bson-objectid'
 import {
   $createParagraphNode,
   $getNodeByKey,
@@ -20,8 +13,6 @@ import {
   $isRangeSelection,
   $isRootOrShadowRoot,
   COMMAND_PRIORITY_EDITOR,
-  COMMAND_PRIORITY_LOW,
-  COPY_COMMAND,
 } from 'lexical'
 import { useEffect, useState } from 'react'
 
@@ -35,11 +26,6 @@ import { $createInlineBlockNode, $isInlineBlockNode } from '../nodes/InlineBlock
 import { INSERT_BLOCK_COMMAND, INSERT_INLINE_BLOCK_COMMAND } from './commands.js'
 
 export type InsertBlockPayload = BlockFieldsOptionalID
-
-type SerializedUnknownLexicalNode = {
-  children?: SerializedUnknownLexicalNode[]
-  type: string
-}
 
 export const BlocksPlugin: PluginComponent = () => {
   const [editor] = useLexicalComposerContext()
@@ -114,59 +100,6 @@ export const BlocksPlugin: PluginComponent = () => {
           return true
         },
         COMMAND_PRIORITY_EDITOR,
-      ),
-      // Remove duplicated ids from clipboard. We do it here because:
-      // 1. Browsers do not allow setting the clipboardData in paste event for security reasons.
-      // 2. If you cut instead of paste, the id will be kept, which is a good thing.
-      editor.registerCommand(
-        COPY_COMMAND,
-        (event) => {
-          copyToClipboard(editor, objectKlassEquals(event, ClipboardEvent) ? event : null)
-            .then(() => {
-              if (!(event instanceof ClipboardEvent) || !event.clipboardData) {
-                throw new Error('No clipboard event')
-              }
-              const lexicalStringified = event.clipboardData.getData('application/x-lexical-editor')
-              if (!lexicalStringified) {
-                return true
-              }
-
-              const lexical = JSON.parse(lexicalStringified) as {
-                nodes: SerializedUnknownLexicalNode[]
-              }
-              const changeIds = (node: SerializedUnknownLexicalNode) => {
-                if (
-                  'fields' in node &&
-                  typeof node.fields === 'object' &&
-                  node.fields !== null &&
-                  'id' in node.fields
-                ) {
-                  node.fields.id = new ObjectID.default().toHexString()
-                } else if ('id' in node) {
-                  node.id = new ObjectID.default().toHexString()
-                }
-
-                if (node.children) {
-                  for (const child of node.children) {
-                    changeIds(child)
-                  }
-                }
-              }
-              for (const node of lexical.nodes) {
-                changeIds(node)
-              }
-              const stringified = JSON.stringify(lexical)
-              event.clipboardData.setData('application/x-lexical-editor', stringified)
-            })
-            .catch((error) => {
-              if (event instanceof ClipboardEvent) {
-                event.clipboardData?.setData('application/x-lexical-editor', '')
-              }
-              throw error
-            })
-          return true
-        },
-        COMMAND_PRIORITY_LOW,
       ),
     )
   }, [editor, setCreatedInlineBlock, targetNodeKey, toggleDrawer])
